refactor: API 경로 주석추가 +  루트 수정 경로 변경
