+++ gusto/src/main/java/com/umc/gusto/domain/route/controller/RouteListController.java
@@ -2,13 +2,11 @@ package com.umc.gusto.domain.route.controller;
 
 
 import com.umc.gusto.domain.route.model.request.RouteListRequest;
-import com.umc.gusto.domain.route.model.request.ModifyRouteRequest;
 import com.umc.gusto.domain.route.model.response.RouteListResponse;
 import com.umc.gusto.domain.route.service.RouteListService;
 import com.umc.gusto.global.auth.model.AuthUser;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
@@ -23,7 +21,10 @@ public class RouteListController {
 
     private final RouteListService routeListService;
 
-    // 루트리스트만 추가
+    /**
+     * 루트 내 식당만 추가 == 루트리스트만 추가
+     * [POST] /routeLists/{routeId}
+     */
     @PostMapping("/{routeId}")
     public ResponseEntity<?> createRouteList
     (@PathVariable Long routeId,
@@ -35,7 +36,10 @@ public class RouteListController {
         return ResponseEntity.ok().build();
     }
 
-    // 루르리스트 항목 삭제
+    /**
+     * 루트 내 식당 삭제 == 루트 내 경로 삭제
+     * [DELETE] /routeLists/{routeId}
+     */
     @DeleteMapping("/{routeListId}")
     public ResponseEntity<?> deleteRouteLis(
             @PathVariable Long routeListId,
@@ -45,7 +49,10 @@ public class RouteListController {
         return ResponseEntity.ok().build();
     }
 
-    // 루트리스트 간 거리 조회
+    /**
+     * 내 루트/그룹 루트 간 거리 조회
+     * [GET] /routeLists/{routeId}/order
+     */
     @GetMapping("/{routeId}/order")
     public ResponseEntity<List<RouteListResponse.RouteList>> getRouteListOrder(
             @PathVariable Long routeId)
@@ -63,10 +70,5 @@ public class RouteListController {
         return ResponseEntity.ok().body(routeListService.getRouteListDetail(routeId,authUSer.getUser(),groupId));
     }
 
-    // 루트 수정
-    @PatchMapping("/{routeId}")
-    public ResponseEntity<?> modifyRoute(@PathVariable Long routeId, @RequestBody ModifyRouteRequest request){
-        routeListService.modifyRouteList(routeId,request);
-        return ResponseEntity.ok(HttpStatus.OK);
-    }
+
 }
