+++ gusto/src/main/java/com/umc/gusto/domain/route/service/RouteListService.java
@@ -1,12 +1,9 @@
 package com.umc.gusto.domain.route.service;
 
 import com.umc.gusto.domain.route.entity.Route;
-import com.umc.gusto.domain.route.model.request.ModifyRouteRequest;
 import com.umc.gusto.domain.route.model.request.RouteListRequest;
-import com.umc.gusto.domain.route.model.request.RouteRequest;
 import com.umc.gusto.domain.route.model.response.RouteListResponse;
 import com.umc.gusto.domain.user.entity.User;
-
 
 import java.util.List;
 
@@ -27,8 +24,5 @@ public interface RouteListService {
 
     // 루트 상세 조회
     RouteListResponse.RouteListResponseDto getRouteListDetail(Long routeId,User user, Long groupId);
-
-    // 루트 상세 수정
-    void modifyRouteList(Long routeId,ModifyRouteRequest request);
 
 }
