+++ gusto/src/main/java/com/umc/gusto/domain/route/service/RouteListServiceImpl.java
@@ -5,13 +5,10 @@ import com.umc.gusto.domain.group.repository.GroupMemberRepository;
 import com.umc.gusto.domain.group.repository.GroupRepository;
 import com.umc.gusto.domain.route.entity.Route;
 import com.umc.gusto.domain.route.entity.RouteList;
-import com.umc.gusto.domain.route.model.request.ModifyRoueListRequest;
-import com.umc.gusto.domain.route.model.request.ModifyRouteRequest;
 import com.umc.gusto.domain.route.model.request.RouteListRequest;
 import com.umc.gusto.domain.route.model.response.RouteListResponse;
 import com.umc.gusto.domain.route.repository.RouteListRepository;
 import com.umc.gusto.domain.route.repository.RouteRepository;
-import com.umc.gusto.domain.store.entity.Store;
 import com.umc.gusto.domain.store.repository.StoreRepository;
 import com.umc.gusto.domain.user.entity.User;
 import com.umc.gusto.global.common.BaseEntity;
@@ -150,43 +147,7 @@ public class RouteListServiceImpl implements RouteListService{
 
     }
 
-    @Transactional
-    @Override
-    public void modifyRouteList(Long routeId, ModifyRouteRequest request) {
-        // 루트리스트 갯수가 6개 이하인지 확인
-        // TODO: 루트 리스트.size()가 null 일때,즉 입력 시 루트리스트가 없을 때
-        if (request.getRouteList().size() >= 7) {
-            throw new GeneralException(Code.ROUTELIST_TO_MANY_REQUEST);
-        }
 
-        // 루트 존재 여부 확인
-        Route route = routeRepository.findRouteByRouteIdAndStatus(routeId, BaseEntity.Status.ACTIVE)
-                .orElseThrow(() -> new GeneralException(Code.ROUTE_NOT_FOUND));
-
-        // 루트 이름 변경 확인
-        if (request.getRouteName() != null) {
-            route.updateRouteName(request.getRouteName());
-        }
-
-        for (ModifyRoueListRequest modifyRoueListRequest : request.getRouteList()) {
-            // 변경된 PK값 확인
-            if (modifyRoueListRequest.getRouteListId() != null) {
-                // 해당 루트리스트 조회
-                RouteList routeList = routeListRepository.findById(modifyRoueListRequest.getRouteListId())
-                        .orElseThrow(() -> new GeneralException(Code.ROUTELIST_NOT_FOUND));
-
-                // 루트리스트 내 수정된 컬럼 값만 업데이트 진행
-                if (modifyRoueListRequest.getOrdinal() != null) {
-                    routeList.updateOrdinal(modifyRoueListRequest.getOrdinal());
-                }
-                if (modifyRoueListRequest.getStoreId() != null) {
-                    Store store = storeRepository.findById(modifyRoueListRequest.getStoreId())
-                            .orElseThrow(() -> new GeneralException(Code.STORE_NOT_FOUND));
-                    routeList.updateStore(store);
-                }
-            }
-        }
-    }
 
 
 }
